Add AssetClass collection decorator system
Each Asset Class can define a static method called decorateCollection

This function will be called any time a collection is created for a
given type of asset.  This lets us build a specific set of convenience
methods for each type of asset.

For the first round I added some methods on the ProjectManifest asset
collections.  These methods let us query packages inside of a portfolio
to find things like which packages provide a given type of component
(e.g. themes) or which packages depend on a given package. (Which
packages depend on react?)

Also added the ability to ask a document collection what types of
documents are in it, or what are the different categories of documents?