Add ability to set data on a data source and save
Added the ability to set data and save.  This will be used in a number
of places, right now it was to be able to make changes to package.json
in projects and save them.

I updated the skypager package.json to not do a git tag when you run npm
version. Will test this out and hopefully it does the lerna publish step
and the npm publish step and commit to github all in one command.